requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.2
pillow==10.0.0
matplotlib==3.7.2
pandas==2.0.3
deepface==0.0.79
# DeepFace optional detector backend (recommended)
mediapipe==0.10.14

# Text & audio analyzers
librosa==0.10.1
soundfile==0.12.1
audioread==3.0.1
scikit-learn==1.3.0

# TensorFlow backend used by DeepFace (recommended for some detector backends)
tensorflow==2.13.0
keras==2.13.1
